requirements.txt
absl-py==1.4.0
astor==0.8.1
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814693874/work
cycler==0.11.0
dataclasses==0.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
scikit-learn==0.24.2
scipy==1.5.4
screeninfo==0.8.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==1.15.0
tensorflow-directml==1.15.8
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.64.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.16.0
zipp==3.6.0